refactored the handleSaveEdit to be a utility function
