+++ src/shared-components/inventory/InventoryForm.tsx
@@ -5,6 +5,7 @@ import deleteEntry from "../../utils.ts/inventory/deleteEntry";
 import { useInventoryForm } from "../../hooks/useInventoryForm";
 import addEntry from "../../utils.ts/inventory/addEntry";
 import updateFormFields from "./updateFormFields";
+import saveEdit from "./saveEdit";
 
 // This pulls the exact row definition from your 'inventory' table
 type InventoryEntry = Database["public"]["Tables"]["inventory"]["Row"];
@@ -137,84 +138,6 @@ export function InventoryForm({
     setEditingCell({ id: entryId, field: fieldName });
     setEditingValue(String(currentValue ?? ""));
   };
-
-  // Called when the user confirms their edit (Enter key or clicking away).
-  // We build the updated row, recalculate derived fields, save to Supabase,
-  // then update our local state so the UI reflects the change immediately.
-
-  const handleSaveEdit = async (entry: InventoryEntry, field: FieldConfig) => {
-    if (!editingCell) return;
-
-    // Convert the typed string into the right type for this field
-    const parsedValue =
-      field.type === "number" // checking if field is a number
-        ? field.formName === "quantity" // check if it's specifically the quantity field
-          ? parseInt(editingValue) || 0 // if so, parse as a whole number e.g 5
-          : parseFloat(editingValue) || 0 // otherwise, parse like a decimal number. e.g $12.99
-        : editingValue; // else just use the data type of string
-
-    // Build a copy of the row with the new value applied. e.g parsedValue
-    const updatedEntry: InventoryEntry = {
-      ...entry,
-      [field.formName]: parsedValue,
-    };
-
-    // If the user edited a field that feeds into our calculations,
-    // recompute the derived fields the same way handleChange does.
-    // *** FUTURE EDITS *** - we could DRY this up by extracting the calculation logic into a separate function since it's used in multiple places now. For now, we'll just keep it here.
-    if (
-      ["cost_per_item", "quantity", "sold_price", "shipping_cost"].includes(
-        field.formName,
-      )
-    ) {
-      const costPer =
-        field.formName === "cost_per_item"
-          ? parseFloat(editingValue) || 0
-          : parseFloat(String(entry.cost_per_item)) || 0;
-      const qty =
-        field.formName === "quantity"
-          ? parseInt(editingValue) || 0
-          : parseInt(String(entry.quantity)) || 0;
-      const listPrice =
-        field.formName === "sold_price"
-          ? parseFloat(editingValue) || 0
-          : parseFloat(String(entry.sold_price)) || 0;
-      const shipping =
-        field.formName === "shipping_cost"
-          ? parseFloat(editingValue) || 0
-          : parseFloat(String(entry.shipping_cost)) || 0;
-
-      updatedEntry.total_cost = Number((costPer * qty).toFixed(2));
-      updatedEntry.total_price_sold = Number((listPrice * qty).toFixed(2));
-      updatedEntry.profit = Number(
-        (
-          updatedEntry.total_price_sold -
-          updatedEntry.total_cost -
-          shipping
-        ).toFixed(2),
-      );
-    }
-
-    // Persist to the database
-    const { error } = await supabase
-      .from("inventory")
-      .update(updatedEntry)
-      .eq("id", entry.id);
-
-    if (error) {
-      console.error("Failed to update entry:", error.message);
-      setEditingCell(null);
-      return;
-    }
-
-    // update our local list so the user sees the change right away
-    setInventoryData((prev) =>
-      prev.map((item) => (item.id === entry.id ? updatedEntry : item)),
-    );
-
-    setEditingCell(null);
-  };
-
   // Called when the user presses Escape — discards the draft and closes the input.
   const handleCancelEdit = () => {
     setEditingCell(null);
@@ -374,11 +297,12 @@ export function InventoryForm({
                               // onBlur fires when the user clicks somewhere else.
                               // We treat that as "done" and save the edit.
                               onBlur={() =>
-                                handleSaveEdit(singleFormEntry, field)
+                                saveEdit(singleFormEntry, field, editingValue, editingCell, setInventoryData, setEditingCell)
+                                // handleSaveEdit(singleFormEntry, field)
                               }
                               onKeyDown={(e) => {
                                 if (e.key === "Enter")
-                                  handleSaveEdit(singleFormEntry, field);
+                                  saveEdit(singleFormEntry, field, editingValue, editingCell, setInventoryData, setEditingCell)
                                 if (e.key === "Escape") handleCancelEdit();
                               }}
                               className="bg-transparent border-b border-sky-500 outline-none text-sm w-full min-w-[60px]"
